backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
passlib[bcrypt]
python-multipart
psycopg2-binary
email-validator
